Add collected data on manager
