+++ src/app/dashboard/manager/clients/page.jsx
@@ -4,7 +4,7 @@ import ManagerLayout from "../managerlayout";
 import RightDetails from "@/components/common/Dashboard/admin/rightDetails";
 import Header from "@/components/common/Dashboard/admin/header";
 import StatsContainer from "@/components/common/Dashboard/admin/statscontainer";
-import TabNav from "../tabNav";
+import TabNav from "./tabNav";
 import Users from "@/components/json/users";
 import Navs from "@/components/json/manager/navbar";
 import UserStats from "@/components/json/manager/userStats";
@@ -19,7 +19,7 @@ const Page = () => {
   return (
     <ManagerLayout>
       <div className="flex justify-between items-start">
-        <div className="w-full lg:w-[75%] xl:w-[80%] px-3">
+        <div className="w-full px-3">
           <Header Navs={Navs} />
           <StatsContainer Stats={UserStats} />
           <TabNav
@@ -28,15 +28,6 @@ const Page = () => {
             setSelectedTr={setSelectedTr}
             handleRowClick={handleRowClick}
           />
-        </div>
-        <div className="lg:w-[25%] xl:w-[20%] h-screen pl-1 bg-fixed bg-white">
-          <div className="lg:w-[20%]  xl:w-[16.5%] h-screen fixed bg-white ">
-            <RightDetails
-              selectedTr={selectedTr}
-              setSelectedTr={setSelectedTr}
-              item={selectedTr !== null ? Users[selectedTr] : null}
-            />
-          </div>
         </div>
       </div>
     </ManagerLayout>
