+++ src/app/dashboard/manager/menus/tabNav.jsx
@@ -97,7 +97,7 @@ const TabNav = ({ Dishes }) => {
         {addUserModal && <AddMenu closeModal={userHandleModal} />}
       </div>
 
-      <div className="w-full grid grid-cols-2 md:grid-cols-3  xl:grid-cols-4">
+      <div className="w-full grid grid-cols-2 md:grid-cols-3 lg:grid-cols-4  xl:grid-cols-5">
         {filteredDishes.map((item, index) => (
           <>
             <AdminCard
